Phase 4: Decompose TrackEditor.tsx (1460 → 452 lines)
Extract VisualEditor, CourseForm, AddCourseDialog, AddTrackDialog into
src/components/track-editor/. Extract useTrackEditorForm hook for form
state and memoized VisualEditor callbacks (eliminates 5x callback
duplication). Add parseSectorLine utility to trackUtils. Consolidate
compact/non-compact rendering to share selectionUI, manageModeContent,
and dialog components.
